+++ Editor/Controls/SplineElementRectSelector.cs
@@ -5,6 +5,14 @@ namespace UnityEditor.Splines
 {
     class SplineElementRectSelector
     {
+        enum Mode
+        {
+            None,
+            Replace,
+            Add,
+            Subtract
+        }
+        
         static class Styles
         {
             public static readonly GUIStyle selectionRect = GUI.skin.FindStyle("selectionRect");
@@ -12,8 +20,11 @@ namespace UnityEditor.Splines
 
         Rect m_Rect;
         Vector2 m_StartPos;
+        Mode m_Mode;
+        Mode m_InitialMode;
         static readonly HashSet<ISplineElement> s_SplineElementsCompareSet = new HashSet<ISplineElement>();
         static readonly List<ISplineElement> s_SplineElementsBuffer = new List<ISplineElement>();
+        static readonly HashSet<ISplineElement> s_PreRectSelectionElements = new HashSet<ISplineElement>();
 
         public void OnGUI(IReadOnlyList<IEditableSpline> paths)
         {
@@ -24,7 +35,31 @@ namespace UnityEditor.Splines
             {
                 case EventType.Layout:
                     if (!Tools.viewToolActive)
+                    {
                         HandleUtility.AddDefaultControl(id);
+
+                        if (m_Mode != Mode.None)
+                        {
+                            // If we've started rect select in Add or Subtract modes, then if we were in a Replace 
+                            // mode just before (i.e. the shift or action has been released temporarily), 
+                            // we need to bring back the pre rect selection elements into current selection.
+                            if (m_InitialMode != Mode.Replace && RefreshSelectionMode())
+                            {
+                                SplineSelection.Clear();
+                                s_SplineElementsCompareSet.Clear();
+
+                                if (m_Mode != Mode.Replace)
+                                {
+                                    foreach (var element in s_PreRectSelectionElements)
+                                        SplineSelection.Add(element);
+                                }
+                                
+                                m_Rect = GetRectFromPoints(m_StartPos, evt.mousePosition);
+                                UpdateSelection(m_Rect, paths);
+                            }
+                        }
+                    }
+
                     break;
 
                 case EventType.Repaint:
@@ -72,10 +107,18 @@ namespace UnityEditor.Splines
 
         protected virtual void BeginSelection(IReadOnlyList<IEditableSpline> paths)
         {
+            RefreshSelectionMode();
+            m_InitialMode = m_Mode;
+            
             s_SplineElementsCompareSet.Clear();
             s_SplineElementsBuffer.Clear();
-            if(!Event.current.shift)
+            if (m_Mode == Mode.Replace)
+            {
                 SplineSelection.Clear();
+                s_PreRectSelectionElements.Clear();
+            }
+            else
+                SplineSelection.GetSelectedElements(s_PreRectSelectionElements);
         }
 
         protected virtual void UpdateSelection(Rect rect, IReadOnlyList<IEditableSpline> paths)
@@ -89,22 +132,54 @@ namespace UnityEditor.Splines
                     GetElementSelection(rect, spline, j, s_SplineElementsBuffer);
             }
 
-            //Compare current frame buffer with last frame's to find new addition
             foreach (var splineElement in s_SplineElementsBuffer)
             {
-                //If wasn't there last frame, add to selection
-                if (!s_SplineElementsCompareSet.Remove(splineElement))
+                //Compare current frame buffer with last frame's to find new additions/removals
+                var wasInRectLastFrame = s_SplineElementsCompareSet.Remove(splineElement);
+                if (m_Mode == Mode.Replace || m_Mode == Mode.Add)
+                {
+                    var canAdd = m_Mode == Mode.Replace ? true : !s_PreRectSelectionElements.Contains(splineElement);
+                    if (!wasInRectLastFrame && canAdd)
+                        SplineSelection.Add(splineElement);
+                } 
+                else if (m_Mode == Mode.Subtract && !wasInRectLastFrame)
+                {
+                    SplineSelection.Remove(splineElement);
+                }
+            }
+            
+            //Remaining spline elements from last frame are removed from selection (or added if mode is subtract)
+            foreach (var splineElement in s_SplineElementsCompareSet)
+            {
+                if (m_Mode == Mode.Replace || m_Mode == Mode.Add)
+                {
+                    // If we're in Add mode, don't remove elements that were in select prior to rect selection 
+                    if (m_Mode == Mode.Add && s_PreRectSelectionElements.Contains(splineElement))
+                        continue;
+                    SplineSelection.Remove(splineElement);
+                }
+                else if (m_Mode == Mode.Subtract && s_PreRectSelectionElements.Contains(splineElement))
                     SplineSelection.Add(splineElement);
             }
-
-            //Remaining spline elements from last frame are removed from selection
-            foreach (var splineElement in s_SplineElementsCompareSet)
-                SplineSelection.Remove(splineElement);
-
+            
             //Move current elements buffer to hash set for next frame compare
             s_SplineElementsCompareSet.Clear();
             foreach (var splineElement in s_SplineElementsBuffer)
                 s_SplineElementsCompareSet.Add(splineElement);
+        }
+
+        bool RefreshSelectionMode()
+        {
+            var modeBefore = m_Mode;
+            if (Event.current.shift)
+                m_Mode = Mode.Add;
+            else if (EditorGUI.actionKey)
+                m_Mode = Mode.Subtract;
+            else
+                m_Mode = Mode.Replace;
+
+            // Return true if the mode has changed
+            return m_Mode != modeBefore;
         }
 
         void GetElementSelection(Rect rect, IEditableSpline spline, int index, List<ISplineElement> results)
@@ -127,6 +202,7 @@ namespace UnityEditor.Splines
 
         protected virtual void EndSelection(Rect rect, IReadOnlyList<IEditableSpline> paths)
         {
+            m_Mode = m_InitialMode = Mode.None;
         }
 
         static Rect GetRectFromPoints(Vector2 a, Vector2 b)