+++ Editor/Utilities/SplineGizmoUtility.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using UnityEditor.SettingsManagement;
 using UnityEngine;
 using UnityEngine.Splines;
@@ -18,30 +17,10 @@ namespace UnityEditor.Splines
         {
             s_GizmosLineColor.value = SettingsGUILayout.SettingsColorField("Splines Color", s_GizmosLineColor, searchContext);
         }
-        
-        const int k_SegmentsCount = 32;
+
         static Vector3[] s_Points;
-
+        static Vector3 s_CameraUp;
         static Color s_OutlineColor = new Color(0f, 0f, 0f, .5f);
-
-        static Dictionary<Spline, List<Vector3>> s_SplineCache = new Dictionary<Spline, List<Vector3>>();
-        static Dictionary<Spline, Vector3[]> s_SplineCacheTable = new Dictionary<Spline, Vector3[]>();
-        static Dictionary<Spline, List<BezierCurve>> s_CurveCache = new Dictionary<Spline, List<BezierCurve>>();
-
-        [InitializeOnLoadMethod]
-        static void Initialize()
-        {
-            EditorSplineUtility.afterSplineWasModified += delegate(Spline spline)
-            {
-                if(s_SplineCache.ContainsKey(spline))
-                    s_SplineCache[spline].Clear();
-                if(s_CurveCache.ContainsKey(spline))
-                {
-                    s_CurveCache[spline].Clear();
-                    CacheSplineCurves(spline);
-                }
-            };
-        }
         
         /// <summary>
         /// Draw a line gizmo for a <see cref="ISplineProvider"/>.
@@ -54,66 +33,58 @@ namespace UnityEditor.Splines
             if (splines == null)
                 return;
 
-            var cameraUp = SceneView.lastActiveSceneView.camera.transform.up;
+            s_CameraUp = SceneView.lastActiveSceneView.camera.transform.up;
             var localToWorld = ((MonoBehaviour)provider).transform.localToWorldMatrix;
             foreach(var spline in splines)
             {
                 if(spline == null || spline.Count < 2)
                     continue;
 
-                if(!s_SplineCacheTable.ContainsKey(spline))
-                    s_SplineCacheTable.Add(spline, null);
-
-                int c = spline.Closed ? spline.Count : spline.Count - 1;
-
-                if(s_SplineCacheTable[spline] == null)
-                {
-                    s_SplineCacheTable[spline] = new Vector3[c * k_SegmentsCount];
-                    CacheCurvePositionsTable(spline, localToWorld, c);
-                }
+                Vector3[] positions;
+                SplineCacheUtility.GetCachedPositions(spline, out positions);
 
                 var color = Gizmos.color;
-                var from = localToWorld.MultiplyPoint(s_SplineCacheTable[spline][0]);
-                for(int i = 1; i < c * k_SegmentsCount; ++i)
+                var from = localToWorld.MultiplyPoint(positions[0]);
+                var previousDir = Vector3.zero;
+                for(int i = 1; i < positions.Length; ++i)
                 {
-                    var to = localToWorld.MultiplyPoint(s_SplineCacheTable[spline][i]);
+                    var to = localToWorld.MultiplyPoint(positions[i]);
                 
                     var center = ( from + to ) / 2f;
                     var size = .1f * HandleUtility.GetHandleSize(center);
                 
                     var dir = to - from;
-                    if(dir.magnitude > size)
+                    var delta = previousDir.magnitude == 0 ? 1f :Vector3.Dot(previousDir, dir.normalized);
+                    //If the angle is too wide between 2 positions, take the previous position to draw the line
+                    if(delta < 0.9f)
                     {
-                        Gizmos.DrawLine(from, to);
-                        Gizmos.color = s_OutlineColor;
-                        //make the gizmo a little thicker
-                        var offset = size * cameraUp / 7.5f;
-                        Gizmos.DrawLine(from - offset, to - offset);
-                
-                        from = to;
                         Gizmos.color = color;
+                        DrawLineSegment(from, from + previousDir, size);
+                        from = from + previousDir;
+                        dir = to - from;
                     }
+                    //Is the second position far enough to draw the segment
+                    if(i == positions.Length-1 || dir.magnitude > size)
+                    {
+                        Gizmos.color = color;
+                        DrawLineSegment(from, to, size);
+                        from = to;
+                        previousDir = Vector3.zero;
+                    }
+                    else
+                        previousDir = dir;
                 }
                 Gizmos.matrix = Matrix4x4.identity; 
             }
         }
 
-        static void CacheSplineCurves(Spline spline)
+        static void DrawLineSegment(Vector3 from, Vector3 to, float size)
         {
-            for(int i = 0, c = spline.Closed ? spline.Count : spline.Count - 1; i < c; ++i)
-                s_CurveCache[spline].Add(spline.GetCurve(i));
-        }
-        
-        static void CacheCurvePositionsTable(Spline spline, Matrix4x4 localToWorld, int curveCount, int segments = k_SegmentsCount)
-        {
-            float inv = 1f / (segments - 1);
-            for(int i = 0; i < curveCount; ++i)
-            {
-                var curve = spline.GetCurve(i);
-                var startIndex = i * k_SegmentsCount;
-                for(int n = 0; n < segments; n++)
-                    (s_SplineCacheTable[spline])[startIndex + n] = CurveUtility.EvaluatePosition(curve, n * inv);
-            }
+            Gizmos.DrawLine(from, to);
+            Gizmos.color = s_OutlineColor;
+            //make the gizmo a little thicker
+            var offset = size * s_CameraUp / 7.5f;
+            Gizmos.DrawLine(from - offset, to - offset);
         }
     }
 }