+++ Runtime/NativeSpline.cs
@@ -62,7 +62,8 @@ namespace UnityEngine.Splines
         /// </summary>
         /// <returns>An IEnumerator that is used to iterate the <see cref="BezierKnot"/> collection.</returns>
         public IEnumerator<BezierKnot> GetEnumerator() => m_Knots.GetEnumerator();
-
+            
+        /// <inheritdoc cref="GetEnumerator"/>
         IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
 
         /// <summary>